Made some changes, Did not use git for a while.
im an idiot. I got preoccupied with trying to understand and use
assembly8086 that I forgot about using version control.

Probably forgot about version control when making the project because
I find it hard to use and collaborate with others, I don't know how
to use branches and share remote repository files.

PURPOSE:
The files removed are of the initial version of the program,
our group was planning to go for a calculator, then after
a recent revision, the project will be a grade calculator.

The new files are files related to the average grade calculator.