Deleted files initial files and revised project
Deleted the files about the project from the initial stage of the
project as a calculator. After a review and revision of the
ongoing status of out project by our subject professor, the
project is changed from a simple calculator to a
averageCalculatorForGrades.

Deleted files from the initial stage of the project, when the project
was about making a calculator for assembly8086.

Added new files for creating a average grade calculator. Most of
which is .asm files because the .exe file will be recompiled by the
emu8086 emulator when the code base changes, and the emu8086 can
compile the program at a high speed, though the compile time
may be inconvenient, but it is better than having to manage multiple
.exe files and other files from building the executable of the program.