Arranged project directory(NO DOCUMENTATION YET)
The project presentation and defense is done. However, I did not
properly document the code. Nor did I update the repository(local
and remote), I know how to add files to the repository and not much
for DELETE, UPDATE, READ operations with git (Git and GitHub). I will
provide further explanations why I did not utilize git and
GitHub down below at the very bottom.

I want to properly document the project and it's progress. Though
the project is done, for future uses and purposes, I want to write
down the reasons about the design decisions and choices that was
made during the development of the project.

REASONS WHY I DID NOT USE GIT AND GITHUB (properly):

I made poor decisions, I rushed things, then I lost focus,
I got overconfident and I was pre-occupied.
The combination of all this resulted to me ignoring
documentation and version control of the project.